+++ source/index.html
@@ -13,7 +13,7 @@
       type="module"
     ></script>
     <script src="assets/scripts/main.js" type="module"></script>
-    <script src="assets/scripts/spoonacular.js" type="module"></script>
+    <script src="assets/scripts/api/spoonacular.js" type="module"></script>
     <script src="assets/scripts/filter.js" type="module"></script>
 
     <!--  Stylesheets  -->
